+++ server/channels/store/storetest/mocks/SessionStore.go
@@ -80,15 +80,15 @@ func (_m *SessionStore) Get(c request.CTX, sessionIDOrToken string) (*model.Sess
 }
 
 // GetLRUSessions provides a mock function with given fields: c, userID, offset
-func (_m *SessionStore) GetLRUSessions(c request.CTX, userID string, offset int) ([]*model.Session, error) {
+func (_m *SessionStore) GetLRUSessions(c request.CTX, userID string, offset uint64) ([]*model.Session, error) {
 	ret := _m.Called(c, userID, offset)
 
 	var r0 []*model.Session
 	var r1 error
-	if rf, ok := ret.Get(0).(func(request.CTX, string, int) ([]*model.Session, error)); ok {
+	if rf, ok := ret.Get(0).(func(request.CTX, string, uint64) ([]*model.Session, error)); ok {
 		return rf(c, userID, offset)
 	}
-	if rf, ok := ret.Get(0).(func(request.CTX, string, int) []*model.Session); ok {
+	if rf, ok := ret.Get(0).(func(request.CTX, string, uint64) []*model.Session); ok {
 		r0 = rf(c, userID, offset)
 	} else {
 		if ret.Get(0) != nil {
@@ -96,7 +96,7 @@ func (_m *SessionStore) GetLRUSessions(c request.CTX, userID string, offset int)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(request.CTX, string, int) error); ok {
+	if rf, ok := ret.Get(1).(func(request.CTX, string, uint64) error); ok {
 		r1 = rf(c, userID, offset)
 	} else {
 		r1 = ret.Error(1)
