+++ server/channels/store/timerlayer/timerlayer.go
@@ -7673,7 +7673,7 @@ func (s *TimerLayerSessionStore) Get(c request.CTX, sessionIDOrToken string) (*m
 	return result, err
 }
 
-func (s *TimerLayerSessionStore) GetLRUSessions(c request.CTX, userID string, offset int) ([]*model.Session, error) {
+func (s *TimerLayerSessionStore) GetLRUSessions(c request.CTX, userID string, offset uint64) ([]*model.Session, error) {
 	start := time.Now()
 
 	result, err := s.SessionStore.GetLRUSessions(c, userID, offset)
